Default to No Selection Criteria for Marker Genes (#130)
Our current methods for selecting marker genes impose some constraints
on what can be considered a true marker gene for a cell type. This
produces a very confusing user experience when for instance, for the
12th out of 12 cell types, there are no valid marker genes. This causes
the pipeline to crash. It might be a more sensible default to just pick
the best available N genes and use those for the plots, and users can
use the more strict criteria later on if they wish.