Handle Anndata Possibly Containing Sparse Arrays (#139)
Now that we use the spatialdata-io package to read the visium data, the
anndata.X attribute will be a sparse array (it used to be dense with our
home-spun visium reader. This creates some problems since all downstream
classes expect a dense numpy array.

Also noticed we had two identical methods to return the count array, so
reduced to one.